Revert "Upgrade project JDK to 25."
This reverts commit d6a7d20a0a5e279e0d0177c295eaa134d1c56d32.
