+++ gradle.properties
@@ -1,7 +1,7 @@
 #
 # Copyright (C) 2024-2026 OpenAni and contributors.
 #
-# 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+# \u6B64\u6E90\u4EE3\u7801\u7684\u4F7F\u7528\u53D7 GNU AFFERO GENERAL PUBLIC LICENSE version 3 \u8BB8\u53EF\u8BC1\u7684\u7EA6\u675F, \u53EF\u4EE5\u5728\u4EE5\u4E0B\u94FE\u63A5\u627E\u5230\u8BE5\u8BB8\u53EF\u8BC1.
 # Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
 #
 # https://github.com/open-ani/ani/blob/main/LICENSE
@@ -27,7 +27,7 @@ android.compile.sdk=36
 android.min.sdk=27
 # JBR must have JCEF bundled
 jvm.toolchain.vendor=jetbrains
-jvm.toolchain.version=25
+jvm.toolchain.version=21
 kotlin.mpp.enableCInteropCommonization=true
 kotlin.apple.xcodeCompatibility.nowarn=true
 kotlin.native.ignoreDisabledTargets=true
